requirements.txt
# Django and database
Django==3.2.25  # LTS version compatible with Djongo
djongo==1.3.6
pymongo==3.12.3
six==1.16.0  # Required by Djongo but not included in newer Django

# REST API
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7

# Authentication
django-allauth==0.58.2

# Internationalization
django-modeltranslation==0.18.11

# Task queue
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1
django-redis==5.4.0  # For Redis cache integration

# Storage
django-storages==1.14.2
Pillow==10.1.0  # Required for ImageField

# Monitoring
django-prometheus==2.3.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
sentry-sdk==1.40.0
python-magic==0.4.27  # For file type validation
